# Fix GPU-CPU tensor manipulation. Small performance boost