There are two fixes:
1. Greedy and random masks are created on CPU side, but are used on both CPU and GPU sides, particularly it selects `logits` from GPU (potential performance reduction). After fix there are two pairs of masks on CPU and `logits` device side, moreover CPU masks are created only if it needs.
2. Replace `res_random = torch.multinomial(probs, 1, True).cpu().numpy()[:, 0]` by `res_random = torch.multinomial(probs, 1, True)[:, 0].cpu().numpy()`. As I understand in the first case we copy the slice from old numpy to new one after copying full tensor from GPU to CPU, but in the second case we get slice view (without memory copying) and copying from GPU to CPU only slice tensor (not full)

Results of benchmark (~**1.6%** performance boost):
```
MISTRAL (python serve/benchmarks/benchmark_throughput.py --local-id mistral-7b-instruct-q0f16-presharded-1gpu --dataset /opt/models/dataset/ShareGPT_V3_unfiltered_cleaned_split.json --seed 0 --num-prompts 1000)

batch-serving:
Engine Throughput: 49.18 requests/s, 18816.81 tokens/s
Engine Throughput: 48.36 requests/s, 18503.69 tokens/s
Engine Throughput: 48.79 requests/s, 18668.08 tokens/s
AVERAGE: 48.78 requests/s, 18662.86 tokens/s

Fix 1:
Engine Throughput: 48.92 requests/s, 18717.29 tokens/s
Engine Throughput: 49.37 requests/s, 18891.61 tokens/s
Engine Throughput: 48.73 requests/s, 18647.23 tokens/s
AVERAGE: 49.01 requests/s, 18752.04 tokens/s

Fix1 + Fix2:
Engine Throughput: 49.23 requests/s, 18837.64 tokens/s
Engine Throughput: 49.43 requests/s, 18911.63 tokens/s
Engine Throughput: 50.00 requests/s, 19130.03 tokens/s
AVERAGE: 49.55 requests/s, 18959.77 tokens/s
```