@masahi 2024-01-30
I think the second optimization is not really helping since the output of `multinomial` has shape `(num_tokens, 1)`. So the number of elements that are copied are the same.